Display error messages if something goes wrong during TPR loading.
git-svn-id: svn+ssh://trac.godzil.net/gcc4ti@365 9552661e-59e3-4036-b4f2-dbe53926924f

+++ trunk/ktigcc/tpr.h
@@ -105,6 +105,6 @@ typedef struct
   TPRFileList quill_files;
 } TPRDataStruct;
 
-short loadTPR(QString &fileName,TPRDataStruct *dest);
+int loadTPR(QString &fileName,TPRDataStruct *dest);
 
 QString loadFileText(const char *fileName);
